Use increasing ports for dev_server fixtures

+++ tests/conftest.py
@@ -20,16 +20,17 @@ import pytest
 from werkzeug import serving
 from werkzeug.utils import cached_property
 from werkzeug._compat import to_bytes
+from itertools import count
 
 
 try:
     __import__('pytest_xprocess')
 except ImportError:
-    @pytest.fixture(scope='session')
+    @pytest.fixture(scope='function')
     def subprocess():
         pytest.skip('pytest-xprocess not installed.')
 else:
-    @pytest.fixture(scope='session')
+    @pytest.fixture(scope='function')
     def subprocess(xprocess):
         return xprocess
 
@@ -64,9 +65,6 @@ def _dev_server():
     serving.run_simple(hostname='localhost', application=app,
                        **testsuite_app.kwargs)
 
-if __name__ == '__main__':
-    _dev_server()
-
 
 class _ServerInfo(object):
     xprocess = None
@@ -95,7 +93,6 @@ class _ServerInfo(object):
             except Exception as e:  # urllib also raises socketerrors
                 print(self.url)
                 print(e)
-        return False
 
     def wait_for_reloader(self):
         old_pid = self.last_pid
@@ -116,7 +113,9 @@ class _ServerInfo(object):
                 return
 
 
-@pytest.fixture
+port_generator = count(13220)
+
+@pytest.fixture(scope='function')
 def dev_server(tmpdir, subprocess, request, monkeypatch):
     '''Run werkzeug.serving.run_simple in its own process.
 
@@ -127,8 +126,9 @@ def dev_server(tmpdir, subprocess, request, monkeypatch):
     def run_dev_server(application):
         app_pkg = tmpdir.mkdir('testsuite_app')
         appfile = app_pkg.join('__init__.py')
+        port = next(port_generator)
         appfile.write('\n\n'.join((
-            'kwargs = dict(port=5001)',
+            'kwargs = dict(port=%d)' % port,
             textwrap.dedent(application)
         )))
 
@@ -151,7 +151,7 @@ def dev_server(tmpdir, subprocess, request, monkeypatch):
 
         def preparefunc(cwd):
             args = [sys.executable, __file__, str(tmpdir)]
-            return info.request_pid, args
+            return lambda: 'pid=%s' % info.request_pid(), args
 
         subprocess.ensure('dev_server', preparefunc, restart=True)
 
@@ -159,10 +159,15 @@ def dev_server(tmpdir, subprocess, request, monkeypatch):
             # Killing the process group that runs the server, not just the
             # parent process attached. xprocess is confused about Werkzeug's
             # reloader and won't help here.
-            pid = info.last_pid
-            os.killpg(os.getpgid(pid), signal.SIGTERM)
+            pid = info.request_pid()
+            if pid:
+                os.killpg(os.getpgid(pid), signal.SIGTERM)
         request.addfinalizer(teardown)
 
         return info
 
     return run_dev_server
+
+
+if __name__ == '__main__':
+    _dev_server()
